src/app/components/admin/manager/manager.component.ts
import { Component, OnDestroy, OnInit } from '@angular/core';
import { Title } from '@angular/platform-browser';
import { AuthService } from '../../../services/auth.service';
import { Subscription } from 'rxjs';

@Component({
  selector: 'app-manager',
  templateUrl: './manager.component.html',
  styleUrls: ['./manager.component.css']
})
export class ManagerComponent implements OnInit, OnDestroy {
  title: string = 'V Queen Management';
  subscription: Subscription | undefined;
    constructor(
    private titleService: Title,
    private authService: AuthService,
  ) { }

  ngOnInit() {
    this.titleService.setTitle('VQAdmin - Main');
  }

  isLoggedIn(): boolean {
    return !!this.authService.accessToken;
  }

  logout(): void {
    this.authService.logout();
  }

  ngOnDestroy(): void {
    if (this.subscription) {
      this.subscription.unsubscribe();
    }
  }

}


